Extract start DM logic to a helper file (#8317)
* Extract start DM logic to a helper file

* Fix incorrect import
